+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Version 1.4 (not yet released)
+-----------
+
+* Changed the implementation of the civicrm_container hook.
+
 Version 1.3
 -----------